requirements.txt
tensorboard==2.3
numpy==1.23.4
scipy
open3d>=0.8
Pillow
tqdm
pyyaml